fix: comments module deleted and merged into bigPictureModule
